backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
httpx==0.27.0
pandas==2.2.2
scikit-learn==1.5.0
groq==0.9.0
pydantic==2.7.1
